BROKEN Sonnet iteration on widget updating
trying to have widget loads trigger only on the widget that has been
edited, instead of the whole dashboard.

The widgets should probably update on a timer anyway, so loading all of
them again isn't _That_ big of a deal.

Dashboards don't have a high number of widgets, and don't scale to
support them well anyway